Conflicts:
	Debug/SDL_Game_Framework.exe
	Debug/SDL_Game_Framework.ilk
	Debug/SDL_Game_Framework.pdb
	SDL_Game_Framework.sdf
	SDL_Game_Framework.v12.suo
	SDL_Game_Framework/Debug/AW_Sprite_Interface.obj
	SDL_Game_Framework/Debug/Character.obj
	SDL_Game_Framework/Debug/Character_Factory_Implementation.obj
	SDL_Game_Framework/Debug/End_Menu_Splash.obj
	SDL_Game_Framework/Debug/Enemy.obj
	SDL_Game_Framework/Debug/Game.obj
	SDL_Game_Framework/Debug/Jesus.obj
	SDL_Game_Framework/Debug/Level.obj
	SDL_Game_Framework/Debug/Level_One.obj
	SDL_Game_Framework/Debug/Main.obj
	SDL_Game_Framework/Debug/NPC_Base.obj
	SDL_Game_Framework/Debug/NPC_Hinter.obj
	SDL_Game_Framework/Debug/NPC_Shopkeeper.obj
	SDL_Game_Framework/Debug/Ogre.obj
	SDL_Game_Framework/Debug/Play_Screen.obj
	SDL_Game_Framework/Debug/Player.obj
	SDL_Game_Framework/Debug/SDL_Game.8E28E0D2.tlog/CL.read.1.tlog
	SDL_Game_Framework/Debug/SDL_Game.8E28E0D2.tlog/CL.write.1.tlog
	SDL_Game_Framework/Debug/SDL_Game.8E28E0D2.tlog/link.read.1.tlog
	SDL_Game_Framework/Debug/SDL_Game_Framework.log
	SDL_Game_Framework/Debug/Splash_Screen.obj
	SDL_Game_Framework/Debug/Start_Menu_Splash.obj
	SDL_Game_Framework/Debug/Test_Everything.obj
	SDL_Game_Framework/Debug/Text_Box.obj
	SDL_Game_Framework/Debug/Text_Boxes.obj
	SDL_Game_Framework/Debug/vc120.idb
	SDL_Game_Framework/Debug/vc120.pdb

+++ SDL_Game_Framework/AW_Sprite_Interface.h
@@ -17,7 +17,7 @@ public:
 	void Render();
 
 	bool Move_By(int xAmount, int yAmount);
-	bool Move_To(int x, int y);
+	bool Move_To(float x, float y);
 
 	bool Move_Increment(int x, int y);
 	bool Move_Between();
